Add Pixel correction to tile-selection
Adding half a pixel to either side of the transform
makes sure that we clamp to a pixel within the tile
and not the outside which may be empty.

Not sure if that's _actually_ what I'm doing here, but
the result is the same.

Tis avoids having seams between the tiles
`