Refactor linked_memory.
We got rid of the allocated tree. Instead of storing the allocated links
in a separate tree we add them to the main-memory space as a header to
the allocated data. Allocated tree could grow really big.

When choosing a size for memory_tree one should not choose the exact
ammount needed. Rather it's usecase is as a for variable size Arena
where we have minimal fragmentation and allocation / frees at
O(Log(fragments)) rather than O(log(Allocations) * Log(Fragments))